fix vehicle late multiplier

+++ wrappers/ortools_vrp_pb.rb
@@ -10,7 +10,7 @@ Google::Protobuf::DescriptorPool.generated_pool.build do
   add_message "ortools_vrp.TimeWindow" do
     optional :start, :int64, 1
     optional :end, :int64, 2
-    optional :late_multiplier, :uint32, 3
+    optional :late_multiplier, :float, 3
   end
   add_message "ortools_vrp.Service" do
     repeated :time_windows, :message, 1, "ortools_vrp.TimeWindow"
